+++ R/data.R
@@ -1,0 +1,13 @@
+#' power_studies_results
+#' 
+#' the results of the included power studies
+#' @format ## 'power_studies_results'
+#' A list of matrices with powers
+"power_studies_results"
+
+#' pvaluecdf
+#' 
+#' data to draw a graph in vignette
+#' @format ## 'pvaluecdf'
+#' A matrix
+"pvaluecdf"