+++ inst/shiny-example/r2sample_shiny/server.R
@@ -3,7 +3,7 @@ library(ggplot2)
 library(Rcpp)
 library(grid)
 library("R2sample")
-#methods <- c("chi large", "chi small", "t test", "KS", "Kuiper", "AD", "CdM", "LR", "ZK", "ZA", "ZC", "Wassp1")
+#methods <- c("chi large", "chi small", "KS", "Kuiper", "AD", "CdM", "LR", "ZK", "ZA", "ZC", "Wassp1")
 
 shinyServer(function(input, output,session) {
 
@@ -26,10 +26,6 @@ shinyServer(function(input, output,session) {
       }
 
     })
-
-
-    
-    
 
     test=eventReactive(input$gobutton, {
         z=indata()